pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-alphazero"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "python-chess>=1.10",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "onnx>=1.14",
    "onnxruntime>=1.15",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.24",
]

[project.optional-dependencies]
dev = ["pytest>=7.3", "pytest-cov>=4.1", "httpx>=0.24"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "web*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
